Improve script.js readability and functionality
Declare most variables and the playRound() function within the
playGame() function now to adhere to the principle of encapsulation.

Added let to variable declarations in getComputerChoice() and
getHumanChoice() to avoid polluting the global scope.

Now getHumanChoice() will repeatedly ask the user for a valid input
instead of defaulting to "rock".